Moved function to context file
Improved description of function
